requirements.txt
fastapi[all]==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy~=2.0.30
python-dotenv~=1.0.1
pydantic~=2.11.4
python-jose~=3.4.0
passlib~=1.7.4
starlette~=0.36.3